pyproject.toml
[tool.poetry]
name = "travel-agency"
version = "0.1.0"
description = "Simple ai trip planning"
authors = ["Dion Baccus"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = {extras = ["tools"], version = "^0.30.11"}
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
sqlalchemy = "^2.0.30"
flask = "^3.0.3"
pytest = "^8.2.2"
alembic = "^1.13.1"
pytest-dotenv = "^0.5.2"
ollama = "^0.2.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"